requirements.txt
flask==3.1.0
flask_sockets==0.2.1
numpy==2.1.3
pandas==2.2.3
xgboost==2.1.2
scikit-learn==1.5.2
openpyxl==3.1.5
tensorboard==2.18.0
torch-geometric==2.6.1